Several fixes

+++ ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: rhutchin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/23 10:22:23 by rhutchin          #+#    #+#             */
-/*   Updated: 2019/05/23 10:37:00 by rhutchin         ###   ########.fr       */
+/*   Updated: 2019/05/28 16:23:25 by rhutchin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(const char *str, int fd)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (str != NULL)
